+++ ocean_data_parser/__init__.py
@@ -2,3 +2,4 @@ import re
 from pathlib import Path
 
 PARSERS = re.findall('parser = "(.*)"', (Path(__file__).parent / "read.py").read_text())
+__version__ = "0.4.0"
