--- ocean_data_parser/_version.py
@@ -1,1 +1,0 @@
-__version__ = "0.3.0"
